+++ R/grpmean.R
@@ -15,7 +15,6 @@
 #'   \code{NULL}, so no weights are used.
 #' @param digits Numeric, amount of digits after decimal point when rounding
 #'   estimates and values.
-#' @param weight.by Deprecated.
 #'
 #' @inheritParams hdi
 #'
@@ -53,20 +52,13 @@
 #' @importFrom sjmisc to_value is_empty
 #' @importFrom rlang enquo .data quo_name
 #' @export
-grpmean <- function(x, dv, grp, weights = NULL, digits = 2, out = c("txt", "viewer", "browser"), weight.by) {
+grpmean <- function(x, dv, grp, weights = NULL, digits = 2, out = c("txt", "viewer", "browser")) {
 
   out <- match.arg(out)
 
   if (out != "txt" && !requireNamespace("sjPlot", quietly = TRUE)) {
     message("Package `sjPlot` needs to be loaded to print HTML tables.")
     out <- "txt"
-  }
-
-  ## TODO remove deprecated argument later
-
-  if (!missing(weight.by)) {
-    message("Argument `weight.by` is deprecated. Please use `weights`.")
-    weights <- weight.by
   }
 
   # create quosures
@@ -79,9 +71,8 @@ grpmean <- function(x, dv, grp, weights = NULL, digits = 2, out = c("txt", "view
     if (inherits(.weights, "try-error")) .weights <- NULL
 
     w.string <- try(eval(weights), silent = TRUE)
-    if (!inherits(w.string, "try-error") && is.character(w.string)) .weights <- w.string
+    if (!inherits(w.string, "try-error") && !is.null(w.string) && is.character(w.string)) .weights <- w.string
 
-    if (!is.null(w.string) && is.character(w.string)) .weights <- w.string
     if (sjmisc::is_empty(.weights) || .weights == "NULL") .weights <- NULL
   } else
     .weights <- NULL