+++ NEWS.md
@@ -12,6 +12,7 @@
 ## Bug fixes
 
 * `model_frame()` now correctly names the offset-columns for terms provided as `offset`-argument (i.e. for models where the offset was not specified inside the formula).
+* Fixed issue with `weights`-argument in `grpmean()` when variable name was passed as character vector.
 
 # sjstats 0.17.1